migrating accessibility to its own namespace
