+++ BlueprintUICommonControls/Sources/AccessibilityElement.swift
@@ -1,7 +1,11 @@
 import BlueprintUI
 import UIKit
 
+
 public struct AccessibilityElement: Element {
+    public typealias Trait = Accessibility.Trait
+    public typealias CustomAction = Accessibility.CustomAction
+    public typealias CustomContent = Accessibility.CustomContent
 
     public var label: String?
     public var value: String?
@@ -214,7 +218,7 @@ extension Element {
         value: String? = nil,
         hint: String? = nil,
         identifier: String? = nil,
-        traits: Set<AccessibilityElement.Trait> = [],
+        traits: Set<Accessibility.Trait> = [],
         accessibilityFrameSize: CGSize? = nil
     ) -> AccessibilityElement {
         accessibilityElement(
@@ -227,202 +231,3 @@ extension Element {
         )
     }
 }
-
-
-
-extension AccessibilityElement {
-    public enum Trait {
-        /// Used in conjunction with UIAccessibilityTrait.adjustable, these will be called to allow accessible adjustment of a value, for example in a slider or stepper control.
-        /// See [Accessibility Increment Documentation](https://developer.apple.com/documentation/objectivec/nsobject/1615076-accessibilityincrement) for further information.
-        public typealias IncrementAction = () -> Void
-
-        /// Used in conjunction with UIAccessibilityTrait.adjustable, these will be called to allow accessible adjustment of a value, for example in a slider or stepper control.
-        /// See [Accessibility Decrement Documentation](https://developer.apple.com/documentation/objectivec/nsobject/1615169-accessibilitydecrement) for further information.
-        public typealias DecrementAction = () -> Void
-
-        case button
-        case link
-        case header
-        case searchField
-        case image
-        case selected
-        case playsSound
-        case keyboardKey
-        case staticText
-        case summaryElement
-        case notEnabled
-        case updatesFrequently
-        case startsMediaSession
-        case adjustable(IncrementAction, DecrementAction)
-        case allowsDirectInteraction
-        case causesPageTurn
-        case tabBar
-        case backButton
-        case toggleButton
-    }
-}
-
-
-extension AccessibilityElement.Trait: Hashable, Equatable {
-    /// This conformance to `Hashable` is provided to allow traits to be included in a `Set`.
-    /// - Important: ⚠️ This implementation does not take equality of associated values on `.adjustable` into account.  ⚠️
-    private var internalValue: Int {
-        switch self {
-        case .button: return 0
-        case .link: return 1
-        case .header: return 2
-        case .searchField: return 3
-        case .image: return 4
-        case .selected: return 5
-        case .playsSound: return 6
-        case .keyboardKey: return 7
-        case .staticText: return 8
-        case .summaryElement: return 9
-        case .notEnabled: return 10
-        case .updatesFrequently: return 11
-        case .startsMediaSession: return 12
-        case .adjustable: return 13
-        case .allowsDirectInteraction: return 14
-        case .causesPageTurn: return 15
-        case .tabBar: return 16
-        case .backButton: return 17
-        case .toggleButton: return 18
-        }
-    }
-
-    public static func == (lhs: AccessibilityElement.Trait, rhs: AccessibilityElement.Trait) -> Bool {
-        lhs.internalValue == rhs.internalValue
-    }
-
-    public func hash(into hasher: inout Hasher) {
-        hasher.combine(internalValue)
-    }
-}
-
-extension UIAccessibilityTraits {
-
-    public init(with set: Set<AccessibilityElement.Trait>) {
-        self.init(rawValue: UIAccessibilityTraits.none.rawValue)
-        for trait in set {
-            switch trait {
-            case .button:
-                formUnion(.button)
-            case .link:
-                formUnion(.link)
-            case .header:
-                formUnion(.header)
-            case .searchField:
-                formUnion(.searchField)
-            case .image:
-                formUnion(.image)
-            case .selected:
-                formUnion(.selected)
-            case .playsSound:
-                formUnion(.playsSound)
-            case .keyboardKey:
-                formUnion(.keyboardKey)
-            case .staticText:
-                formUnion(.staticText)
-            case .summaryElement:
-                formUnion(.summaryElement)
-            case .notEnabled:
-                formUnion(.notEnabled)
-            case .updatesFrequently:
-                formUnion(.updatesFrequently)
-            case .startsMediaSession:
-                formUnion(.startsMediaSession)
-            case .adjustable:
-                formUnion(.adjustable)
-            case .allowsDirectInteraction:
-                formUnion(.allowsDirectInteraction)
-            case .causesPageTurn:
-                formUnion(.causesPageTurn)
-            case .tabBar:
-                formUnion(.tabBar)
-            case .backButton:
-                formUnion(.backButton)
-            case .toggleButton:
-                formUnion(._toggleButton)
-            }
-        }
-    }
-}
-
-
-extension AccessibilityElement {
-    /// Used to provide additional functionality to assistive technologies beyond your accessible UI.
-    public struct CustomAction: Equatable, Hashable {
-        public typealias OnActivation = () -> Bool
-
-        /// A localized name that discribes the action.
-        public var name: String
-        /// An image representing the action to be shown with some assistive technologies such as Switch Control.
-        public var image: UIImage?
-        /// A Callback for when the action is activated. This should return a `bool` indicating success or failure of the action.
-        public var onActivation: OnActivation
-
-        public init(name: String, image: UIImage? = nil, onActivation: @escaping OnActivation) {
-            self.name = name
-            self.image = image
-            self.onActivation = onActivation
-        }
-
-        public static func == (lhs: AccessibilityElement.CustomAction, rhs: AccessibilityElement.CustomAction) -> Bool {
-            // Disregard onActivation for equatablity pruposes.
-            lhs.name == rhs.name && lhs.image == rhs.image
-        }
-
-        public func hash(into hasher: inout Hasher) {
-            // Disregard onActivation for hash pruposes.
-            hasher.combine(name)
-            hasher.combine(image)
-        }
-    }
-}
-
-extension AccessibilityElement {
-    public struct CustomContent: Equatable {
-        /// The importance of the content.
-        public enum Importance: Equatable {
-            /// By default custom content is available through the rotor.
-            case `default`
-            /// In addtion to being available through the rotor, high importance content will announced in the main VoiceOver utterance.
-            /// High Importance content is announced follllowing the `accessibilityValue` but preceding any `accessibilityHint`.
-            case high
-        }
-
-        public var label: String
-        public var value: String?
-        public var importance: Importance
-
-        public init(label: String, value: String? = nil, importance: Importance = .default) {
-            self.label = label
-            self.value = value
-            self.importance = importance
-        }
-
-        public var axCustomContent: AXCustomContent {
-            let importance: AXCustomContent.Importance
-            switch self.importance {
-            case .high:
-                importance = .high
-            case .default:
-                importance = .default
-            }
-            return .init(label: label, value: value, importance: importance)
-        }
-    }
-}
-
-
-extension AXCustomContent {
-    public convenience init(_ content: AccessibilityElement.CustomContent) {
-        self.init(label: content.label, value: content.value, importance: content.importance == .high ? .high : .default)
-    }
-
-    public convenience init(label: String, value: String?, importance: AXCustomContent.Importance = .default) {
-        self.init(label: label, value: value ?? "")
-        self.importance = importance
-    }
-}
-
